+++ src/hooks/useChunkedOrderLines.js
@@ -1,6 +1,6 @@
 import { useChunkedCQLFetch } from '@folio/stripes/core';
 
-import { ORDER_LINES_ENDPOINT } from '../constants/endpoints';
+import { ORDER_LINES_ENDPOINT } from '../constants';
 // When fetching from a potentially large list of order lines,
 // make sure to chunk the request to avoid hitting limits.
 const useChunkedOrderLines = (poLineIdsArray, queryOptions = {}) => {