+++ src/routes/AgreementLineViewRoute/AgreementLineViewRoute.js
@@ -10,7 +10,7 @@ import View from '../../components/views/AgreementLine';
 import { urls } from '../../components/utilities';
 
 import { useAgreementsHelperApp, useChunkedOrderLines, useSuppressFromDiscovery } from '../../hooks';
-import { AGREEMENT_ENDPOINT, AGREEMENT_LINE_ENDPOINT } from '../../constants/endpoints';
+import { AGREEMENT_ENDPOINT, AGREEMENT_LINE_ENDPOINT } from '../../constants';
 
 const AgreementLineViewRoute = ({
   handlers,